requirements.txt
python-decouple==3.8
requests==2.32.3
python-telegram-bot==21.3
numpy==1.24.4
opencv-python==4.10.0.84
ultralytics==8.2.48
fastapi==0.111.0
uvicorn==0.30.1
# zlib-state==0.1.6
# sudo apt-get install zlib1g-dev